+++ mesh/adj_table.h
@@ -71,7 +71,7 @@ int extract_nodeset_from_sideset(const int                                elemen
                                  ptrdiff_t                               *n_nodes_out,
                                  idx_t **SFEM_RESTRICT                    nodes_out);
 
-int extract_nodeset_from_sidesets(uint16_t                                 n_sidesets,
+int extract_nodeset_from_sidesets(ptrdiff_t                                n_sidesets,
                                   const enum ElemType                      element_type[],
                                   idx_t **const SFEM_RESTRICT              elems[],
                                   const ptrdiff_t                          n_surf_elements[],
